Removes unneeded export default from transaction-state, implements within create-identity-form and removes unneeded code. Delete transaction-status
